Restructure F2 docs and tests
