+++ features/F2/acceptance_tests/acceptance.py
@@ -1,0 +1,23 @@
+from .s1 import f2s1
+from .s2 import f2s2
+from .s3 import f2s3
+from .s4 import f2s4
+from .s5 import f2s5
+from .s6 import f2s6
+from .s7 import f2s7
+from .s8 import f2s8
+from .s9 import f2s9
+from .s10 import f2s10
+
+__all__ = [
+    "f2s1",
+    "f2s2",
+    "f2s3",
+    "f2s4",
+    "f2s5",
+    "f2s6",
+    "f2s7",
+    "f2s8",
+    "f2s9",
+    "f2s10",
+]
