+++ features/F2/acceptance_tests/s3.py
@@ -1,0 +1,22 @@
+from __future__ import annotations
+
+from pathlib import Path
+
+from shared import compose, compose_paths, dump_logs, search_meili
+
+from .helpers import _run_once
+
+
+def f2s3(tmp_path: Path) -> None:
+    compose_file, workdir, output_dir = compose_paths(__file__)
+    try:
+        _run_once(compose_file, workdir, output_dir)
+        docs = search_meili(compose_file, workdir, "copies = 1")
+        assert all(doc["copies"] == 1 for doc in docs)
+    except Exception:
+        dump_logs(compose_file, workdir)
+        raise
+    finally:
+        compose(
+            compose_file, workdir, "down", "--volumes", "--rmi", "local", check=False
+        )
